progressing on plex config
